commit final

+++ api/routes/apiRouter.js
@@ -1,10 +1,6 @@
 const express = require('express')
 const routerAPI = express.Router()
-
-
-routerAPI.use(express.urlencoded({ extended: true })) // processar o body
-routerAPI.use(express.json()) // processa o body em formato json
-
+const jwt = require('jsonwebtoken')
 const knex = require('knex')({
     client: 'pg',
     debug: true,
@@ -14,8 +10,59 @@ const knex = require('knex')({
     }
 });
 
+routerAPI.use(express.urlencoded({ extended: true })) // processar o body
+routerAPI.use(express.json()) // processa o body em formato json
+
+
+// Checa se possui autenticação
+let checkToken = (req, res, next) => {
+    let authToken = req.headers["authorization"]
+    if (!authToken) {
+        res.status(401).json({ message: 'Token de acesso requerida' })
+    }
+    else {
+        let token = authToken.split(' ')[1]
+        req.token = token
+    }
+
+    jwt.verify(req.token, process.env.SECRET_KEY, (err, decodeToken) => {
+        if (err) {
+            res.status(401).json({ message: 'Acesso negado' })
+            return
+        }
+        req.usuarioId = decodeToken.id
+        next()
+    })
+}
+
+// Verifica se a autenticação possui perfil ADMIN
+let isAdmin = (req, res, next) => {
+    knex
+        .select('*').from('usuario').where({ id: req.usuarioId })
+        .then((usuarios) => {
+            if (usuarios.length) {
+                let usuario = usuarios[0]
+                let roles = usuario.roles.split(';')
+                let adminRole = roles.find(i => i === 'ADMIN')
+                if (adminRole === 'ADMIN') {
+                    next()
+                    return
+                }
+                else {
+                    res.status(403).json({ message: 'Role de ADMIN requerida' })
+                    return
+                }
+            }
+        })
+        .catch(err => {
+            res.status(500).json({
+                message: 'Erro ao verificar roles de usuário - ' + err.message
+            })
+        })
+}
+
 // Obter a lista de produtos 
-routerAPI.get('/produtos', (req, res, next) => {
+routerAPI.get('/produtos', checkToken, (req, res, next) => {
     //res.json(lista_produtos)
     knex.select('*').from('produto')
         .then(data => res.status(200).json(data))
@@ -27,7 +74,7 @@ routerAPI.get('/produtos', (req, res, next) => {
 })
 
 // Obter um produto específico 
-routerAPI.get('/produtos/:id', (req, res, next) => {
+routerAPI.get('/produtos/:id', checkToken, (req, res, next) => {
     let id = parseInt(req.params.id)
 
     knex.select('*')
@@ -51,7 +98,7 @@ routerAPI.get('/produtos/:id', (req, res, next) => {
 })
 
 // Incluir um produto 
-routerAPI.post('/produtos', (req, res, next) => {
+routerAPI.post('/produtos', checkToken, isAdmin, (req, res, next) => {
     console.log(req.body)
     if (Object.keys(req.body).length === 0) {
         res.status(400).json({ message: 'Produto deve ser passado no corpo da requisição para o cadastro' })
@@ -74,7 +121,7 @@ routerAPI.post('/produtos', (req, res, next) => {
 })
 
 // Alterar um produto
-routerAPI.put('/produtos/:id', (req, res, next) => {
+routerAPI.put('/produtos/:id', checkToken, isAdmin, (req, res, next) => {
     console.log(req.body)
     let id = parseInt(req.params.id)
     if (Object.keys(req.body).length === 0) {
@@ -82,15 +129,15 @@ routerAPI.put('/produtos/:id', (req, res, next) => {
     }
     else if (id === 0) {
         res.status(400).json({ message: 'id deve ser maior que 0' })
-    }  
+    }
     else {
-        knex.where({id: id})
-            .update({descricao: req.body.descricao, valor: req.body.valor, marca: req.body.marca})
+        knex.where({ id: id })
+            .update({ descricao: req.body.descricao, valor: req.body.valor, marca: req.body.marca })
             .table('produto')
             .then(data => {
                 console.log(data)
-                res.status(!!data?200:404).json({
-                    message: !!data?'Produto atualizado com sucesso' : 'Produto não encontrado',
+                res.status(!!data ? 200 : 404).json({
+                    message: !!data ? 'Produto atualizado com sucesso' : 'Produto não encontrado',
                     updatedRows: data
                 })
             })
@@ -103,20 +150,20 @@ routerAPI.put('/produtos/:id', (req, res, next) => {
 })
 
 // Excluir um produto 
-routerAPI.delete('/produtos/:id', (req, res, next) => {
+routerAPI.delete('/produtos/:id', checkToken, isAdmin, (req, res, next) => {
     console.log(req.body)
     let id = parseInt(req.params.id)
     if (id === 0) {
         res.status(400).json({ message: 'id deve ser maior que 0' })
-    }  
+    }
     else {
-        knex.where({id: id})
+        knex.where({ id: id })
             .delete()
             .table('produto')
             .then(data => {
                 console.log(data)
-                res.status(!!data?200:404).json({
-                    message: !!data?'Produto excluído com sucesso' : 'Produto não encontrado',
+                res.status(!!data ? 200 : 404).json({
+                    message: !!data ? 'Produto excluído com sucesso' : 'Produto não encontrado',
                     updatedRows: data
                 })
             })
